// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const AccountType = {
  CASH: 'CASH',
  CHECKING: 'CHECKING',
  DEPOSIT: 'DEPOSIT',
  SAVINGS: 'SAVINGS',
  CREDIT_CARD: 'CREDIT_CARD'
} as const

export type AccountType = (typeof AccountType)[keyof typeof AccountType]


export const CategoryType = {
  INCOME: 'INCOME',
  EXPENSE: 'EXPENSE'
} as const

export type CategoryType = (typeof CategoryType)[keyof typeof CategoryType]


export const TransactionType = {
  INCOME: 'INCOME',
  EXPENSE: 'EXPENSE',
  TRANSFER: 'TRANSFER'
} as const

export type TransactionType = (typeof TransactionType)[keyof typeof TransactionType]


export const PeriodType = {
  MONTHLY: 'MONTHLY',
  WEEKLY: 'WEEKLY',
  CUSTOM: 'CUSTOM'
} as const

export type PeriodType = (typeof PeriodType)[keyof typeof PeriodType]


export const SourceType = {
  MANUAL: 'MANUAL',
  CSV: 'CSV',
  MAIL: 'MAIL'
} as const

export type SourceType = (typeof SourceType)[keyof typeof SourceType]


export const MatchType = {
  CONTAINS: 'CONTAINS',
  EXACT: 'EXACT',
  STARTS_WITH: 'STARTS_WITH',
  ENDS_WITH: 'ENDS_WITH'
} as const

export type MatchType = (typeof MatchType)[keyof typeof MatchType]


export const BankName = {
  BANCO_DE_CHILE: 'BANCO_DE_CHILE',
  BANCO_SANTANDER: 'BANCO_SANTANDER',
  BANCO_ESTADO: 'BANCO_ESTADO',
  SCOTIABANK: 'SCOTIABANK',
  BCI: 'BCI',
  ITAU: 'ITAU',
  BANCO_FALABELLA: 'BANCO_FALABELLA',
  BANCO_RIPLEY: 'BANCO_RIPLEY',
  BANCO_CONSORCIO: 'BANCO_CONSORCIO',
  BANCO_SECURITY: 'BANCO_SECURITY',
  BANCO_BICE: 'BANCO_BICE',
  BANCO_INTERNACIONAL: 'BANCO_INTERNACIONAL',
  COOPEUCH: 'COOPEUCH',
  TENPO: 'TENPO',
  MERCADO_PAGO: 'MERCADO_PAGO',
  OTRO: 'OTRO'
} as const

export type BankName = (typeof BankName)[keyof typeof BankName]